+++ crates/core/config/tests/workspace_test.rs
@@ -1,6 +1,6 @@
 use moon_config::{
-    ConfigError, GeneratorConfig, HasherConfig, NodeConfig, NotifierConfig, RunnerConfig,
-    VcsConfig, VcsManager, WorkspaceConfig, WorkspaceProjects,
+    ConfigError, GeneratorConfig, HasherConfig, NotifierConfig, RunnerConfig, VcsConfig,
+    VcsManager, WorkspaceConfig, WorkspaceProjects,
 };
 use moon_constants::CONFIG_WORKSPACE_FILENAME;
 use moon_utils::test::get_fixtures_dir;
@@ -31,10 +31,8 @@ fn loads_defaults() {
                 generator: GeneratorConfig::default(),
                 extends: None,
                 hasher: HasherConfig::default(),
-                node: None,
                 notifier: NotifierConfig::default(),
                 projects: WorkspaceProjects::default(),
-                typescript: None,
                 vcs: VcsConfig::default(),
                 schema: String::new(),
             }
@@ -46,7 +44,6 @@ fn loads_defaults() {
 
 mod extends {
     use super::*;
-    use moon_config::{NodePackageManager, TypeScriptConfig, YarnConfig};
     use pretty_assertions::assert_eq;
     use std::fs;
 
@@ -63,39 +60,12 @@ mod extends {
                     log_running_command: false,
                     ..RunnerConfig::default()
                 },
-                node: Some(NodeConfig {
-                    version: "4.5.6".into(),
-                    add_engines_constraint: true,
-                    dedupe_on_lockfile_change: false,
-                    package_manager: NodePackageManager::Yarn,
-                    yarn: Some(YarnConfig {
-                        plugins: None,
-                        version: "3.0.0".into()
-                    }),
-                    ..NodeConfig::default()
-                }),
                 vcs: VcsConfig {
                     manager: VcsManager::Svn,
                     ..VcsConfig::default()
                 },
                 ..WorkspaceConfig::default()
             }
-        );
-    }
-
-    #[test]
-    fn recursive_merges_typescript() {
-        let fixture = get_fixtures_dir("config-extends/workspace");
-        let config = WorkspaceConfig::load(fixture.join("typescript-2.yml")).unwrap();
-
-        assert_eq!(
-            config.typescript,
-            Some(TypeScriptConfig {
-                create_missing_config: false,
-                root_config_file_name: "tsconfig.root.json".into(),
-                sync_project_references: true,
-                ..TypeScriptConfig::default()
-            })
         );
     }
 
@@ -175,491 +145,6 @@ mod extends {
             )?;
 
             super::load_jailed_config(jail.directory())?;
-
-            Ok(())
-        });
-    }
-
-    #[test]
-    fn loads_from_file() {
-        figment::Jail::expect_with(|jail| {
-            fs::create_dir_all(jail.directory().join("shared")).unwrap();
-
-            jail.create_file(
-                format!("shared/{}", super::CONFIG_WORKSPACE_FILENAME),
-                include_str!("../../../../tests/fixtures/config-extends/.moon/workspace.yml"),
-            )?;
-
-            jail.create_file(
-                super::CONFIG_WORKSPACE_FILENAME,
-                r#"
-extends: ./shared/workspace.yml
-
-node:
-    version: '18.0.0'
-    npm:
-        version: '8.0.0'
-"#,
-            )?;
-
-            let config: WorkspaceConfig = super::load_jailed_config(jail.directory())?;
-
-            // Inherits from extended file
-            assert!(!config.node.as_ref().unwrap().add_engines_constraint);
-            assert!(!config.typescript.unwrap().sync_project_references);
-            assert_eq!(config.vcs.manager, VcsManager::Svn);
-
-            // Ensure we can override the extended config
-            assert_eq!(config.node.as_ref().unwrap().version, "18.0.0".to_owned());
-            assert_eq!(
-                config.node.as_ref().unwrap().npm.version,
-                "8.0.0".to_owned()
-            );
-
-            Ok(())
-        });
-    }
-
-    #[test]
-    fn loads_from_url() {
-        figment::Jail::expect_with(|jail| {
-            jail.set_env(
-                "MOON_WORKSPACE_ROOT",
-                jail.directory().to_owned().to_string_lossy(),
-            );
-
-            jail.create_file(
-                    super::CONFIG_WORKSPACE_FILENAME,
-r#"
-extends: https://raw.githubusercontent.com/moonrepo/moon/master/tests/fixtures/config-extends/.moon/workspace.yml
-
-node:
-    version: '18.0.0'
-    npm:
-        version: '8.0.0'
-"#,
-                )?;
-
-            let config: WorkspaceConfig = super::load_jailed_config(jail.directory())?;
-
-            // Inherits from extended file
-            assert!(!config.node.as_ref().unwrap().add_engines_constraint);
-            assert!(!config.typescript.unwrap().sync_project_references);
-            assert_eq!(config.vcs.manager, VcsManager::Svn);
-
-            // Ensure we can override the extended config
-            assert_eq!(config.node.as_ref().unwrap().version, "18.0.0".to_owned());
-            assert_eq!(
-                config.node.as_ref().unwrap().npm.version,
-                "8.0.0".to_owned()
-            );
-
-            Ok(())
-        });
-    }
-
-    // #[test]
-    // #[should_panic(expected = "TODO")]
-    // fn handles_invalid_url() {
-    //     figment::Jail::expect_with(|jail| {
-    //         jail.create_file(
-    //             super::CONFIG_WORKSPACE_FILENAME,
-    //             "extends: https://raw.githubusercontent.com/this/is/an/invalid/file.yml",
-    //         )?;
-
-    //         super::load_jailed_config(jail.directory())?;
-
-    //         Ok(())
-    //     });
-    // }
-}
-
-mod node {
-    use super::*;
-    use moon_config::NodePackageManager;
-
-    #[test]
-    fn loads_defaults() {
-        figment::Jail::expect_with(|jail| {
-            jail.create_file(
-                CONFIG_WORKSPACE_FILENAME,
-                r#"
-projects: {}
-node:
-    packageManager: yarn"#,
-            )?;
-
-            let config = super::load_jailed_config(jail.directory())?;
-
-            assert_eq!(
-                config,
-                WorkspaceConfig {
-                    node: Some(NodeConfig {
-                        package_manager: NodePackageManager::Yarn,
-                        ..NodeConfig::default()
-                    }),
-                    projects: WorkspaceProjects::default(),
-                    ..WorkspaceConfig::default()
-                }
-            );
-
-            Ok(())
-        });
-    }
-
-    #[test]
-    #[should_panic(
-        expected = "invalid type: found unsigned int `123`, expected struct NodeConfig for key \"workspace.node\""
-    )]
-    fn invalid_type() {
-        figment::Jail::expect_with(|jail| {
-            jail.create_file(super::CONFIG_WORKSPACE_FILENAME, "node: 123")?;
-
-            super::load_jailed_config(jail.directory())?;
-
-            Ok(())
-        });
-    }
-
-    #[test]
-    #[should_panic(
-        expected = "Must be a valid semantic version for key \"workspace.node.version\""
-    )]
-    fn invalid_version() {
-        figment::Jail::expect_with(|jail| {
-            jail.create_file(
-                super::CONFIG_WORKSPACE_FILENAME,
-                r#"
-node:
-  version: 'foo bar'
-projects:
-  foo: packages/foo"#,
-            )?;
-
-            super::load_jailed_config(jail.directory())?;
-
-            Ok(())
-        });
-    }
-
-    #[test]
-    #[should_panic(
-        expected = "Must be a valid semantic version for key \"workspace.node.version\""
-    )]
-    fn no_patch_version() {
-        figment::Jail::expect_with(|jail| {
-            jail.create_file(
-                super::CONFIG_WORKSPACE_FILENAME,
-                r#"
-node:
-  version: '16.13'
-projects:
-  foo: packages/foo"#,
-            )?;
-
-            super::load_jailed_config(jail.directory())?;
-
-            Ok(())
-        });
-    }
-
-    #[test]
-    #[should_panic(
-        expected = "Must be a valid semantic version for key \"workspace.node.version\""
-    )]
-    fn no_minor_version() {
-        figment::Jail::expect_with(|jail| {
-            jail.create_file(
-                super::CONFIG_WORKSPACE_FILENAME,
-                r#"
-node:
-  version: '16'
-projects:
-  foo: packages/foo"#,
-            )?;
-
-            super::load_jailed_config(jail.directory())?;
-
-            Ok(())
-        });
-    }
-
-    #[test]
-    #[should_panic(
-        expected = "unknown variant: found `what`, expected `one of `npm`, `pnpm`, `yarn`` for key \"workspace.node.packageManager\""
-    )]
-    fn invalid_package_manager() {
-        figment::Jail::expect_with(|jail| {
-            jail.create_file(
-                super::CONFIG_WORKSPACE_FILENAME,
-                r#"
-node:
-  version: '16.13.0'
-  packageManager: what
-projects:
-  foo: packages/foo"#,
-            )?;
-
-            super::load_jailed_config(jail.directory())?;
-
-            Ok(())
-        });
-    }
-
-    #[test]
-    fn valid_package_manager() {
-        figment::Jail::expect_with(|jail| {
-            jail.create_file(
-                super::CONFIG_WORKSPACE_FILENAME,
-                r#"
-node:
-  version: '16.13.0'
-  packageManager: yarn
-projects:
-  foo: packages/foo"#,
-            )?;
-
-            super::load_jailed_config(jail.directory())?;
-
-            Ok(())
-        });
-    }
-
-    #[test]
-    fn inherits_from_env_var() {
-        figment::Jail::expect_with(|jail| {
-            jail.set_env("MOON_NODE_VERSION", "4.5.6");
-
-            jail.create_file(
-                super::CONFIG_WORKSPACE_FILENAME,
-                r#"
-node:
-    version: '16.13.0'
-projects: {}
-"#,
-            )?;
-
-            let config = super::load_jailed_config(jail.directory())?;
-
-            assert_eq!(config.node.unwrap().version, String::from("4.5.6"));
-
-            Ok(())
-        });
-    }
-}
-
-mod npm {
-    #[test]
-    #[should_panic(
-        expected = "invalid type: found string \"foo\", expected struct NpmConfig for key \"workspace.node.npm\""
-    )]
-    fn invalid_type() {
-        figment::Jail::expect_with(|jail| {
-            jail.create_file(
-                super::CONFIG_WORKSPACE_FILENAME,
-                r#"
-node:
-    version: '16.13.0'
-    npm: foo"#,
-            )?;
-
-            super::load_jailed_config(jail.directory())?;
-
-            Ok(())
-        });
-    }
-
-    #[test]
-    #[should_panic(
-        expected = "Must be a valid semantic version for key \"workspace.node.npm.version\""
-    )]
-    fn invalid_version() {
-        figment::Jail::expect_with(|jail| {
-            jail.create_file(
-                super::CONFIG_WORKSPACE_FILENAME,
-                r#"
-node:
-    version: '16.13.0'
-    npm:
-        version: 'foo bar'
-projects:
-  foo: packages/foo
-"#,
-            )?;
-
-            super::load_jailed_config(jail.directory())?;
-
-            Ok(())
-        });
-    }
-
-    #[test]
-    fn inherits_from_env_var() {
-        figment::Jail::expect_with(|jail| {
-            jail.set_env("MOON_NPM_VERSION", "4.5.6");
-
-            jail.create_file(
-                super::CONFIG_WORKSPACE_FILENAME,
-                r#"
-node:
-    version: '16.13.0'
-    npm:
-        version: '1.2.3'
-projects: {}
-"#,
-            )?;
-
-            let config = super::load_jailed_config(jail.directory())?;
-
-            assert_eq!(config.node.unwrap().npm.version, String::from("4.5.6"));
-
-            Ok(())
-        });
-    }
-}
-
-mod pnpm {
-
-    #[test]
-    #[should_panic(
-        expected = "invalid type: found string \"foo\", expected struct PnpmConfig for key \"workspace.node.pnpm\""
-    )]
-    fn invalid_type() {
-        figment::Jail::expect_with(|jail| {
-            jail.create_file(
-                super::CONFIG_WORKSPACE_FILENAME,
-                r#"
-node:
-    version: '16.13.0'
-    pnpm: foo"#,
-            )?;
-
-            super::load_jailed_config(jail.directory())?;
-
-            Ok(())
-        });
-    }
-
-    #[test]
-    #[should_panic(
-        expected = "Must be a valid semantic version for key \"workspace.node.pnpm.version\""
-    )]
-    fn invalid_version() {
-        figment::Jail::expect_with(|jail| {
-            jail.create_file(
-                super::CONFIG_WORKSPACE_FILENAME,
-                r#"
-node:
-    version: '16.13.0'
-    pnpm:
-        version: 'foo bar'
-projects:
-  foo: packages/foo"#,
-            )?;
-
-            super::load_jailed_config(jail.directory())?;
-
-            Ok(())
-        });
-    }
-
-    #[test]
-    fn inherits_from_env_var() {
-        figment::Jail::expect_with(|jail| {
-            jail.set_env("MOON_PNPM_VERSION", "4.5.6");
-
-            jail.create_file(
-                super::CONFIG_WORKSPACE_FILENAME,
-                r#"
-node:
-    version: '16.13.0'
-    packageManager: 'pnpm'
-    pnpm:
-        version: '1.2.3'
-projects: {}
-"#,
-            )?;
-
-            let config = super::load_jailed_config(jail.directory())?;
-
-            assert_eq!(
-                config.node.unwrap().pnpm.unwrap().version,
-                String::from("4.5.6")
-            );
-
-            Ok(())
-        });
-    }
-}
-
-mod yarn {
-
-    #[test]
-    #[should_panic(
-        expected = "invalid type: found string \"foo\", expected struct YarnConfig for key \"workspace.node.yarn\""
-    )]
-    fn invalid_type() {
-        figment::Jail::expect_with(|jail| {
-            jail.create_file(
-                super::CONFIG_WORKSPACE_FILENAME,
-                r#"
-node:
-    version: '16.13.0'
-    yarn: foo"#,
-            )?;
-
-            super::load_jailed_config(jail.directory())?;
-
-            Ok(())
-        });
-    }
-
-    #[test]
-    #[should_panic(
-        expected = "Must be a valid semantic version for key \"workspace.node.yarn.version\""
-    )]
-    fn invalid_version() {
-        figment::Jail::expect_with(|jail| {
-            jail.create_file(
-                super::CONFIG_WORKSPACE_FILENAME,
-                r#"
-node:
-    version: '16.13.0'
-    yarn:
-        version: 'foo bar'
-projects:
-  foo: packages/foo"#,
-            )?;
-
-            super::load_jailed_config(jail.directory())?;
-
-            Ok(())
-        });
-    }
-
-    #[test]
-    fn inherits_from_env_var() {
-        figment::Jail::expect_with(|jail| {
-            jail.set_env("MOON_YARN_VERSION", "4.5.6");
-
-            jail.create_file(
-                super::CONFIG_WORKSPACE_FILENAME,
-                r#"
-node:
-    version: '16.13.0'
-    packageManager: 'yarn'
-    yarn:
-        version: '1.2.3'
-projects: {}
-"#,
-            )?;
-
-            let config = super::load_jailed_config(jail.directory())?;
-
-            assert_eq!(
-                config.node.unwrap().yarn.unwrap().version,
-                String::from("4.5.6")
-            );
 
             Ok(())
         });
