+++ tests/fixtures/config-extends/.moon/toolchain.yml
@@ -1,0 +1,10 @@
+# NOTE: This is used to test config extending. Changing this will definitely break something!
+
+node:
+  version: '16.1.0'
+  addEnginesConstraint: false
+  npm:
+    version: '7.0.0'
+
+typescript:
+  syncProjectReferences: false
