+++ tests/fixtures/config-extends/toolchain/base-0.yml
@@ -1,0 +1,3 @@
+node:
+  version: '1.2.3'
+  addEnginesConstraint: true
