+++ tests/fixtures/config-extends/toolchain/base-1.yml
@@ -1,0 +1,6 @@
+extends: './base-0.yml'
+
+node:
+  packageManager: 'yarn'
+  yarn:
+    version: '3.0.0'
