+++ tests/fixtures/config-extends/toolchain/base-2.yml
@@ -1,0 +1,5 @@
+extends: './base-1.yml'
+
+node:
+  version: '4.5.6'
+  dedupeOnLockfileChange: false
