+++ tests/fixtures/config-extends/workspace/base-0.yml
@@ -1,6 +1,2 @@
 runner:
   cacheLifetime: '3 hours'
-
-node:
-  version: '1.2.3'
-  addEnginesConstraint: true
