+++ tests/fixtures/config-extends/workspace/base-2.yml
@@ -2,7 +2,3 @@ extends: './base-1.yml'
 
 runner:
   logRunningCommand: false
-
-node:
-  version: '4.5.6'
-  dedupeOnLockfileChange: false
